Split out log extraction utilities from Chrome/Dom code.
